requirements.txt
asgiref==3.4.1
boto3==1.19.9
botocore==1.22.9
Django==3.2.7
django-crispy-forms==1.13.0
django-filter==21.1
django-storages==1.12.3
djangorestframework==3.12.4
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.3.2
psycopg2==2.9.1
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
